Moved own position handling to LocationsService
