+++ app/src/main/java/com/example/neighbourproject/NeighbourApplication.kt
@@ -1,6 +1,8 @@
 package com.example.neighbourproject
 
 import android.app.Application
+import com.example.neighbourproject.location.LocationRepository
+import com.example.neighbourproject.location.LocationService
 import com.example.neighbourproject.neighbour.NeighboursRepository
 import com.example.neighbourproject.neighbour.NeighboursRepositoryTest
 import com.example.neighbourproject.neighbour.NeighboursService
@@ -13,10 +15,12 @@ import org.koin.dsl.module
 class NeighbourApplication: Application() {
     private val appModule = module {
         single<NeighboursService> { NeighboursRepository() }
+        single<LocationService> { LocationRepository() }
     }
 
     private val appModuleTest = module {
         single<NeighboursService> { NeighboursRepositoryTest() }
+        single<LocationService> { LocationRepository() }
     }
 
     override fun onCreate() {
