+++ app/src/main/java/com/example/neighbourproject/location/LocationService.kt
@@ -1,0 +1,9 @@
+package com.example.neighbourproject.location
+
+import com.example.neighbourproject.neighbour.data.Position
+
+interface LocationService {
+    fun setLastPosition(position: Position)
+    fun getLastPosition(): Position?
+    fun calculateDistanceToMyPosition(position: Position): Double
+}
