+++ app/src/main/java/com/example/neighbourproject/neighbour/NeighboursRepositoryTest.kt
@@ -165,22 +165,4 @@ class NeighboursRepositoryTest() : NeighboursService {
     override fun isSignedIn(): Boolean {
         return userProfileRemote.value != null
     }
-
-    private var myPosition : Position? = null
-
-    override fun setLastPosition(position: Position){
-        myPosition = position
-    }
-    override fun getLastPosition(): Position?{
-        return myPosition
-    }
-
-    override fun calculateDistanceToMyPosition(position: Position): Double{
-        myPosition?.let {
-            val results = FloatArray(1)
-            Location.distanceBetween(
-                it.latitude, it.longitude, position.latitude, position.longitude, results)
-            return results[0].toDouble()
-        }?: return -1.0
-    }
 }
