+++ app/src/main/java/com/example/neighbourproject/ui/neigbour/NeighbourViewModel.kt
@@ -1,31 +1,30 @@
 package com.example.neighbourproject.ui.neigbour
 
-import android.location.Location
 import android.util.Log
 import androidx.lifecycle.ViewModel
+import com.example.neighbourproject.location.LocationService
 import com.example.neighbourproject.neighbour.NeighboursService
 import com.example.neighbourproject.neighbour.data.People
 import com.example.neighbourproject.neighbour.data.Position
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import kotlin.math.round
-import kotlin.math.roundToInt
 
 class NeighbourViewModel : ViewModel(), KoinComponent {
     companion object{
         private const val TAG = "NeighbourViewModel"
     }
 
-    private val service: NeighboursService by inject()
+    private val neighboursService: NeighboursService by inject()
+    private val locationService: LocationService by inject()
 
     private var people : People? = null
 
     fun setLastPosition(position: Position){
-        service.setLastPosition(position)
+        locationService.setLastPosition(position)
     }
 
     fun calculateDistanceToMe(position: Position): String{
-        val dist : Double = service.calculateDistanceToMyPosition(position)/1000
+        val dist : Double = locationService.calculateDistanceToMyPosition(position)/1000
         return if(dist < 0)
             "You have no position"
         else
@@ -33,7 +32,7 @@ class NeighbourViewModel : ViewModel(), KoinComponent {
     }
 
     fun selectedNeighbour(id : String): Boolean{
-        people = service.getNeighbourById(id)
+        people = neighboursService.getNeighbourById(id)
         Log.d("NeighbourViewModel", "Setting neighbour: $people")
         return people != null
     }
